requirements.txt
asgiref==3.5.2
Deprecated==1.2.14
dj-rest-auth==3.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-naomi==0.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-extra-fields==3.5.0
factory-boy==3.2.1
Faker==17.0.0
filetype==1.2.0
flake8==6.0.0
jaconv==0.3.4
mccabe==0.7.0
Pillow==9.5.0
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pykakasi==2.2.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
wrapt==1.15.0
